examples/sales_with_leads/requirements.txt
cartesia-line
google-genai>=1.26.0; python_version>='3.9'
python-dotenv>=1.0.0
loguru>=0.7.0
aiohttp>=3.12.0
